[수정] 코드정리
