+++ UJeongApp/LocationSettingPlatform/LocationSettingView.swift
@@ -1,5 +1,5 @@
 //
-//  LocationSetting.swift
+//  LocationSettingView.swift
 //  UJeongApp
 //
 //  Created by Hamlit Jason on 2023/01/03.
@@ -78,8 +78,6 @@ public struct LocationSettingView: View {
                     }
                     .listStyle(.sidebar)
                 }
-                
-                
                 .onAppear {
                     viewStore.send(.onAppear)
                 }
